Skeleton for in-game menu
 * Localized dropbox and jQuery libraries
 * Removed unnecessary plugins
 * Added ability to patch console.log, console.info, etc. to show up in
a specified section of the menu
 * Modified preventUIActions (now configureUIActions) and the
pause/resume functionality on blur to accommodate the flexibility the
in-game menu requires (disabling preventUIActions and preventing
pause/resume on blur/focus)
 * Fixed bug in rgbmode that caused the transitionend callback for some
elements not to occur. Fix isn’t the most elegant way (TODO: check if
children elements have a transitionend event listener, and trigger it).